quantity change type input to int

+++ src/components/inventory/ReceiveStockModal.tsx
@@ -203,7 +203,7 @@ const ReceiveStockModal: React.FC<ReceiveStockModalProps> = ({
                   type="number"
                   required
                   value={formData.quantity}
-                  onChange={(e) => handleInputChange('quantity', parseFloat(e.target.value) || 1)}
+                  onChange={(e) => handleInputChange('quantity', parseInt(e.target.value) || 1)}
                   className="block w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-transparent"
                   placeholder="Quantity"
                 />
