+++ src/components/inventory/TransferStockModal.tsx
@@ -254,7 +254,7 @@ const TransferStockModal: React.FC<TransferStockModalProps> = ({
                 type="number"
                 required
                 value={formData.quantity}
-                onChange={(e) => handleInputChange('quantity', parseFloat(e.target.value))}
+                onChange={(e) => handleInputChange('quantity', parseInt(e.target.value))}
                 className="block w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-purple-500 focus:border-transparent"
                 placeholder="Quantity to transfer"
               />
